requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
seaborn==0.13.2
matplotlib==3.9.2
catboost==1.2.7
xgboost==2.1.3
flask==3.1.0
flask_cors==5.0.0
pyyaml==6.0.2
# -e .